python3files/requirements.txt
numpy==1.23.1
pillow==10.2.0
opencv-python==4.9.0.80
mediapipe==0.10.11
roslibpy==1.8.1
transforms3d==0.4.2